+++ src/Builder.fs
@@ -27,18 +27,20 @@ type internal DefinitionParts = {
 
 [<RequireQualifiedAccess>]
 module internal DefinitionParts =
-    let defaults = {
-        Name = ApplicationName (Name "Console Application")
-        Version = None
-        Title = None
-        ApplicationInfo = ApplicationInfo.Hidden
-        ApplicationOptions = Commands.applicationOptions
-        Output = Output.console
-        Ask = MF.ConsoleStyle.Console.ask
-        Commands = Map.empty
-        DefaultCommand = CommandName (Name CommandNames.List)
-        OptionDecorationLevel = Minimal
-    }
+    let defaults =
+        let output = Output.defaults
+        {
+            Name = ApplicationName (Name "Console Application")
+            Version = None
+            Title = None
+            ApplicationInfo = ApplicationInfo.Hidden
+            ApplicationOptions = Commands.applicationOptions
+            Output = output
+            Ask = output.Ask
+            Commands = Map.empty
+            DefaultCommand = CommandName (Name CommandNames.List)
+            OptionDecorationLevel = Minimal
+        }
 
     let output { Output = output } = output
 
@@ -134,7 +136,7 @@ type ConsoleApplicationBuilder<'r> internal (buildApplication: Definition -> 'r)
 
     [<CustomOperation("useOutput")>]
     member __.UseOutput(state, output): Definition =
-        state <!> fun parts -> { parts with Output = output }
+        state <!> fun parts -> { parts with Output = output; Ask = output.Ask }
 
     [<CustomOperation("useAsk")>]
     member __.UseAsk(state, ask): Definition =